Assignment-3/requirements.txt
flask==3.1.2
werkzeug==3.1.4

# Deep Learning & NLP
tensorflow==2.16.2
transformers==4.40.0
torch==2.3.1

# Image & Audio Processing
pillow==10.1.0
pydub==0.25.1
SpeechRecognition==3.10.1
gtts==2.4.0

# Numerical Computing
numpy==1.26.4

# Tokenization (for translation models)
sentencepiece==0.2.1
